feat: Added sectors implementation and finished experts implementation

+++ server/src/main/kotlin/it/polito/wa2/g13/server/ticketing/experts/ExpertController.kt
@@ -1,12 +1,13 @@
 package it.polito.wa2.g13.server.ticketing.experts
 
+
 import jakarta.validation.Valid
 import org.springframework.validation.BindingResult
 import org.springframework.web.bind.annotation.*
 
 @RestController
 class ExpertController(
-    private val expertService: ExpertService
+    private val expertService: ExpertService,
 ) {
 
     @PostMapping("/API/experts")
@@ -14,7 +15,8 @@ class ExpertController(
                   br: BindingResult): Boolean {
 
         if (!br.hasErrors()) {
-            return if(expertService.setExpert(expertDTO)) true
+            return if(expertService.setExpert(expertDTO))
+                true
             else
                 throw DuplicateExpertException()
         }
@@ -25,9 +27,7 @@ class ExpertController(
 
     @GetMapping("/API/experts/{id}")
     fun getExpertById(@PathVariable id: Long): ExpertDTO? {
-
         return expertService.getExpertById(id) ?: throw ExpertNotFoundException()
-
     }
 
     @PutMapping("/API/experts/{id}")
@@ -50,16 +50,18 @@ class ExpertController(
 
     }
 
-    @GetMapping("/API/experts/?sector={sector}")
-    fun getExpertsBySector(@PathVariable sector: String) : List<ExpertDTO>?{
-
-        return expertService.getExpertsBySector(sector) ?: throw ExpertsOfSelectedSectorNotFoundException()
-
+    @GetMapping("/API/experts/")
+    fun getExpertsBySector(@RequestParam sectorName: String) : List<ExpertDTO>?{
+        return expertService.getExpertsBySector(sectorName) ?: throw ExpertsOfSelectedSectorNotFoundException()
     }
 
     @DeleteMapping("/API/experts/{id}")
     fun deleteExpertById(@PathVariable id: Long) {
+        if (expertService.getExpertById(id)!= null){
+            return expertService.deleteExpertById(id)
+        }else{
+            throw ExpertNotFoundException()
+        }
 
-        return expertService.deleteExpertById(id)
     }
 }
